# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express API for salve health care clinics",
    "version": "1.0.0",
    "URL": "locahost:3000"
  },
  "paths": {
    "/clinics": {
      "get": {
        "summary": "Retrieves a list of all current clinics",
        "description": "This endpoint returns a list of all the clinics currently registered in the system.",
        "responses": {
          "200": {
            "description": "Successfully retrieved a list of clinics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Clinic"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch clinics data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{clinicId}": {
      "get": {
        "summary": "Gets a list of patients for a specified clinic",
        "description": "Retrieves a list of patients associated with the clinic ID provided in the path.",
        "parameters": [
          {
            "in": "path",
            "name": "clinicId",
            "required": true,
            "description": "The ID of the clinic to retrieve patients for.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of patients for the clinic.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Patient"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Patients data for this clinic is not available.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong with your request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Clinic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "established": {
            "type": "string",
            "format": "date",
            "description": "The date when the clinic was established."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "clinic_id": {
            "type": "integer"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date"
          }
        }
      }
    }
  },
  "tags": []
}